pyproject.toml
[project]
name = "geoguessr-locate"
version = "0.1.0"
description = "CLI that uses Google Gemini Vision to guess rough location from GeoGuessr screenshots"
readme = "README.md"
authors = [{name = "Calum"}]
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["geoguessr", "geolocation", "vision", "gemini", "google"]

dependencies = [
  "google-generativeai>=0.7.0",
  "pillow>=10.4.0",
  "requests>=2.32.0",
  "pydantic>=2.7.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0",
  "filelock>=3.15.0",
  "orjson>=3.10.0",
  "folium>=0.16.0"
]

[project.scripts]
geoguessr-locate = "geoguessr_locate.cli:app"
geoguessr-locate-gui = "geoguessr_locate.gui:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

